requirements.txt
gunicorn==21.2.0
dash==3.3.0
plotly==6.5.0
Flask==3.1.2
Werkzeug==3.1.3

pandas==2.3.3
numpy==2.2.6
scipy==1.15.3
scikit-learn==1.7.2

psycopg2-binary==2.9.11

requests==2.32.5
websocket-client==1.9.0
upstox-python-sdk==2.19.0

python-dotenv==1.2.1
pytz==2025.2
tzdata==2025.2

beautifulsoup4==4.14.2
lxml==6.0.2
openpyxl==3.1.5

joblib==1.5.2
tqdm==4.67.1
retrying==1.4.2
typing_extensions==4.15.0
